backend/testing/pipecat-quickstart/pyproject.toml
[project]
name = "pipecat-quickstart"
version = "0.1.0"
description = "Quickstart example for building voice AI bots with Pipecat"
requires-python = ">=3.10"
dependencies = [
    "pipecat-ai[cartesia,daily,deepgram,gemini,local-smart-turn-v3,openai,runner,silero,webrtc,google]",
    "pipecat-ai-cli",
    "strands-agents",
    "litellm",
    "strands-agents-tools",
]

[dependency-groups]
dev = [
    "pyright>=1.1.404,<2",
    "ruff>=0.12.11,<1",
]

[tool.ruff]
line-length = 100
[tool.ruff.lint]
select = ["I"]
